# openapi.yaml
openapi: 3.0.3
info:
  title: Budget Management API
  version: 1.0.0
  description: >
    A comprehensive backend API for managing budgets, expenses, orders, transactions, notifications, and user authentication/management.
  contact:
    name: Son Nguyen
    url: https://example.com
    email: son.nguyen@example.com
servers:
  - url: https://budget-management-backend-api.onrender.com
    description: Production Server
...
paths:
...
  /api/tasks:
    post:
      tags:
        - Tasks
      summary: Create a new task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description]
              properties:
                description:
                  type: string
      responses:
        '201':
          description: Task submitted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        '400':
          description: Validation error.
        '401':
          description: Unauthorized.
        '500':
          description: Server error.

    get:
      tags:
        - Tasks
      summary: Retrieve all tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        '401':
          description: Unauthorized.
        '500':
          description: Server error.
  /api/tasks/{id}:
    get:
      tags:
        - Tasks
      summary: Check the status of a task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        '404':
          description: Task not found.
        '401':
          description: Unauthorized.
        '500':
          description: Server error.

    delete:
      tags:
        - Tasks
      summary: Delete a task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task deleted.
        '404':
          description: Task not found.
        '401':
          description: Unauthorized.
        '500':
          description: Server error.

    put:
      tags:
        - Tasks
      summary: Update a task by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                status:
                  type: string
                  enum: [pending, completed]
      responses:
        '200':
          description: Task updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        '404':
          description: Task not found.
        '401':
          description: Unauthorized.
        '500':
          description: Server error.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        _id:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, completed]
...